requirements.txt
Flask
pymongo
pyJWT
chardet
flask_caching
pandas
scikit-learn
requests
